Add `ParkShim`
ParkShim is a mechanism that injects a closure-shim which takes the
Park timeout, does some work, and then returns a new Park timeout which
tokio will actually use. This is useful for sharing run time across
multiple event loops, since the Park duration is guaranteed by tokio to
otherwise be an idle period.